refactor: shared의 Public API로만 import 하게끔 변경
